fix(st-spi): cast the number of banks from DT to uint8_t
In the DT, we will never have a number of OSPI banks that overflow an
uint8_t (255). This is then OK to cast the return of fdt32_to_cpu() to
uint8_t.

This was found with -Wconversion warning:
drivers/st/spi/stm32_ospi.c:1345:24: warning: conversion from 'uint32_t'
 {aka 'unsigned int'} to 'uint8_t' {aka 'unsigned char'} may change
 value [-Wconversion]
 1345 |                 bank = fdt32_to_cpu(*cuint);
      |                        ^~~~~~~~~~~~

Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/tf-a/+/557401
Tested-by: Yann GAUTIER <yann.gautier@foss.st.com>
ACI: CIBUILD <MDG-smet-aci-builds@list.st.com>
Domain-Review: Christophe KERELLO <christophe.kerello@st.com>

+++ drivers/st/spi/stm32_ospi.c
@@ -1302,7 +1302,7 @@ int stm32_ospi_init(void)
 		}
 
 		for (i = 0U; i < _OMM_MAX_OSPI; i++) {
-			bank = fdt32_to_cpu(*cuint);
+			bank = (uint8_t)fdt32_to_cpu(*cuint);
 			if ((bank >= _OMM_MAX_OSPI) ||
 			    ((bank_assigned & BIT(bank)) != 0U)) {
 				return -FDT_ERR_BADVALUE;
@@ -1342,7 +1342,7 @@ int stm32_ospi_init(void)
 			return -FDT_ERR_BADVALUE;
 		}
 
-		bank = fdt32_to_cpu(*cuint);
+		bank = (uint8_t)fdt32_to_cpu(*cuint);
 		if (bank >= _OMM_MAX_OSPI) {
 			return -FDT_ERR_BADVALUE;
 		}
